8 way directional movement (no attacking)
it can go from Idle 2 to all 8 walking and Running directions. It follows idle > walking > Running. All changes are applied in Demo Level V2
Programmers: I suggest you make a lead in in the movement from idle to walking. For example, when i press forward on my keyboard, it takes 1 second to reach maximum speed. There are 'lead in' animations in case you have sprinting in this game, which would be good to implement sprinting.
With the new blend space, the Attacking animations are no longer being called. You may update the blend space and re add them.