edge detection shader v2 applied to character FOV
- hooked echo actors and logic for sphere mask
- changed mapping to bloom effect
- used same color logic for both shader
- edge shader applied to character FOV, echo shader reverted to previous shader

Due to a math problem in the edge detection shader, it cannot be applied on top of itself, as it basically reverts itself with every consequential trigger. Therefore, the edge detection shader has been applied to the constant ring of view for the player as it creates no problems whatsoever.