Smart Seeking and Predicting Behaviors Pt 2
When a (melee) enemy loses sight of the player, they try to predict where they went. Then, it looks around in the vicinity, in nearby corners and blind spots.
Prediction is a guess of where the player is based on the location and velocity. Seeking is a search of possible spots the player might be hiding.

Seeking:
Whenever an actor exits sight for enough time, the enemy will remember where it was the last time it was seen. Then, activate an EQS and check around corners.
- added a seeking state
- handled a new sense (lack of sight)
- seeking state uses a different system than prediction, getting actual vector location of where it was last seen
- after the enemy goes to the predicted location, it will start seeking
- seeking runs an EQS to check nearby corners and points where the player could have gone realistically

Prediction:
It has been reworked and simplified, as seeking does what predicting was trying to achieve.
- prediction event is called constantly when in sight
- prediction state is called after losing sight
- after trying to move to last known location, seeking behavior is activated

Known issues:
- seeking makes the enemy stand still if no suitable EQS location is found
- prediction makes the enemy stand still if predicted location is unreachable
- dashing or walking into a wall when the enemy is predicting makes it so the prediction location is spawned out of bounds and unreachable